* Added voters * Added new waypoints to script * Number of bug fixes and adjustments

+++ utility/CourseMath.h
@@ -2,9 +2,12 @@
 #define __COURSEMATH_H__
 
 
+#include <stdint.h>
+
+
 class CourseMath {
 public:
-	static double calculateBTW(double gpsLon, double gpsLat, double waypointLon, double waypointLat);
+	static int16_t calculateBTW(double gpsLon, double gpsLat, double waypointLon, double waypointLat);
 	static double calculateDTW(double gpsLon, double gpsLat, double waypointLon, double waypointLat);
 private:
 	CourseMath() { };
